feat(ember-file-upload-mirage): rename export from `upload` to `uploadHandler`
