+++ test-app/tests/unit/upload-file-test.js
@@ -1,7 +1,7 @@
 import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
 import setupMirage from 'ember-cli-mirage/test-support/setup-mirage';
-import { upload as uploadHandler } from 'ember-file-upload-mirage';
+import { uploadHandler } from 'ember-file-upload-mirage';
 import UploadFile from 'ember-file-upload/upload-file';
 import { FileSource } from 'ember-file-upload/interfaces';
